feat: rename utils/prime_config.py to utils/prime.py
Also reword the auth-gate message to suggest `prime login` first.
