+++ verifiers/v1/cli/init.py
@@ -6,7 +6,7 @@ from pathlib import Path
 from pydantic_config import cli
 
 from verifiers.v1.configs.cli.init import InitConfig
-from verifiers.v1.utils.prime_config import ensure_prime_auth
+from verifiers.v1.utils.prime import ensure_prime_auth
 
 USAGE = (
     "usage: uv run init <name> [--path ./environments] [-T/--add-tool] "
