+++ verifiers/v1/utils/platform.py
@@ -17,7 +17,7 @@ import httpx
 from verifiers.v1.configs.cli.eval import EvalConfig
 from verifiers.v1.episode import Episode
 from verifiers.v1.trace import Trace
-from verifiers.v1.utils.prime_config import load_prime_config
+from verifiers.v1.utils.prime import load_prime_config
 
 logger = logging.getLogger(__name__)
