+++ verifiers/v1/utils/prime.py
@@ -21,9 +21,9 @@ def load_prime_config() -> dict:
 
 
 def ensure_prime_auth() -> None:
-    """Exit when no Prime API key is configured (`$PRIME_API_KEY` or `prime login`)."""
+    """Exit when no Prime API key is configured (`prime login` or `$PRIME_API_KEY`)."""
     if os.getenv("PRIME_API_KEY") or load_prime_config().get("api_key"):
         return
     raise SystemExit(
-        "not authenticated with Prime Intellect - set $PRIME_API_KEY or run `prime login`"
+        "not authenticated with prime - run `prime login` or set $PRIME_API_KEY"
     )
